backend/requirements.txt
absl-py==2.3.1
aider-chat==0.86.2
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
aiosqlite==0.22.1
alembic==1.13.1
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
asgiref==3.11.1
async-timeout==4.0.3
asyncpg==0.31.0
attrs==25.4.0
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.14.3
blinker==1.9.0
build==1.4.0
cachetools==6.2.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
chroma-hnswlib==0.7.3
chromadb==1.5.5
click==8.3.1
colorama==0.4.6
coloredlogs==15.0.1
ConfigArgParse==1.7.1
contourpy==1.3.2
coverage==7.13.5
cryptography==46.0.6
cycler==0.12.1
dataclasses-json==0.6.7
diff-match-patch==20241021
diskcache==5.6.3
distro==1.9.0
dnspython==2.8.0
durationpy==0.10
email-validator==2.3.0
exceptiongroup==1.3.1
faiss-cpu==1.13.2
fastapi==0.135.2
fastapi-cli==0.0.24
fastuuid==0.14.0
filelock==3.20.3
filetype==1.2.0
flake8==7.3.0
Flask==3.1.3
flatbuffers==25.12.19
fonttools==4.59.2
frozenlist==1.8.0
fsspec==2026.2.0
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.49.1
google-auth-httplib2==0.2.1
google-genai==1.69.0
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
greenlet==3.3.2
grep-ast==0.9.0
grpcio==1.74.0
grpcio-status==1.71.2
h11==0.16.0
h2==4.3.0
h5py==3.14.0
hf-xet==1.4.3
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.0
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.8.0
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
importlib_metadata==7.2.1
importlib_resources==6.5.2
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.2
json5==0.13.0
jsonpatch==1.33
jsonpointer==3.1.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
keras==3.11.3
kiwisolver==1.4.9
kubernetes==35.0.0
langchain==1.2.13
langchain-chroma==1.1.0
langchain-classic==1.0.3
langchain-community==0.4.1
langchain-core==1.2.23
langchain-google-genai==4.2.1
langchain-huggingface==1.2.1
langchain-nvidia-ai-endpoints==1.2.1
langchain-openai==1.1.11
langchain-qdrant==1.1.0
langchain-text-splitters==1.1.1
langgraph==1.1.3
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.12
langsmith==0.7.23
libclang==18.1.1
litellm==1.81.10
Mako==1.3.10
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==3.26.2
matplotlib==3.10.6
mccabe==0.7.0
mdurl==0.1.2
mixpanel==5.0.0
ml_dtypes==0.5.3
mmh3==5.2.1
mpmath==1.3.0
mslex==1.3.0
multidict==6.7.1
mypy_extensions==1.1.0
namex==0.1.0
narwhals==2.12.0
networkx==3.4.2
numpy==2.2.6
oauthlib==3.3.1
onnxruntime==1.23.2
openai==2.20.0
opendataloader-pdf==2.3.0
opentelemetry-api==1.40.0
opentelemetry-exporter-otlp-proto-common==1.40.0
opentelemetry-exporter-otlp-proto-grpc==1.40.0
opentelemetry-instrumentation==0.61b0
opentelemetry-instrumentation-asgi==0.61b0
opentelemetry-instrumentation-fastapi==0.61b0
opentelemetry-proto==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
opentelemetry-util-http==0.61b0
opt_einsum==3.4.0
optree==0.17.0
orjson==3.11.8
ormsgpack==1.12.2
oslex==0.1.3
overrides==7.7.0
packaging==25.0
pandas==2.3.2
pathspec==1.0.4
pdf2image==1.17.0
pdfminer.six==20251230
pdfplumber==0.11.9
pexpect==4.9.0
pillow==11.3.0
pluggy==1.6.0
portalocker==3.2.0
posthog==7.8.6
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.2.2
ptyprocess==0.7.0
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.3
pycodestyle==2.14.0
pycparser==3.0
pydantic==2.12.4
pydantic-settings==2.13.1
pydantic_core==2.41.5
pydeck==0.9.1
pydub==0.25.1
pyflakes==3.4.0
Pygments==2.19.2
pypandoc==1.16.2
pyparsing==3.2.4
pypdf==6.9.1
PyPDF2==3.0.1
pypdfium2==5.6.0
pyperclip==1.11.0
PyPika==0.51.1
pyproject_hooks==1.2.0
pyreadline3==3.5.4
pytesseract==0.3.13
pytest==8.4.2
pytest-cov==5.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
pytz==2025.2
pywin32==311
PyYAML==6.0.3
qdrant-client==1.17.1
rank-bm25==0.2.2
referencing==0.37.0
regex==2026.1.15
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.1.0
rich-toolkit==0.19.7
rpds-py==0.29.0
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.13.2
sentence-transformers==5.3.0
shellingham==1.5.4
shtab==1.8.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
socksio==1.0.0
sounddevice==0.5.5
soundfile==0.13.1
soupsieve==2.8.3
SQLAlchemy==2.0.48
starlette==1.0.0
streamlit==1.51.0
sympy==1.14.0
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
termcolor==3.1.0
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.2
toml==0.10.2
tomli==2.4.0
torch==2.11.0
tornado==6.5.2
tqdm==4.67.1
transformers==5.4.0
tree-sitter==0.25.2
tree-sitter-c-sharp==0.23.1
tree-sitter-embedded-template==0.25.0
tree-sitter-language-pack==0.13.0
tree-sitter-yaml==0.7.2
typer==0.23.0
typer-slim==0.23.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ujson==5.12.0
uritemplate==4.2.0
urllib3==2.5.0
uuid_utils==0.14.1
uv==0.11.2
uvicorn==0.42.0
watchdog==6.0.0
watchfiles==1.1.1
wcwidth==0.6.0
websocket-client==1.9.0
websockets==16.0
Werkzeug==3.1.3
wrapt==1.17.3
xxhash==3.6.0
yarl==1.23.0
zipp==3.23.0
zstandard==0.25.0
